pyproject.toml
[project]
name = "private-chef"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alibabacloud-oss-v2>=1.2.5",
    "fastapi>=0.136.1",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.46.0",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
